Refactor test_get_settings_default function to use _NssSettings class
This commit refactors the test_get_settings_default function in the test_settings.py file to use the _NssSettings class from the nukeserversocket.settings module. The function now asserts that the settings object is an instance of _NssSettings and that its data matches the defaults. Additionally, the test now checks that the 'temp_value' key is set to True in the settings object's data and that it is present in the file specified by the settings object's path.